+++ src/cosmic_io.c
@@ -1,8 +1,8 @@
 #include "cosmic/cosmic_io.h"
 
 struct cosmic_io_vtable {
-  int (*read)(cosmic_io_t *, char *, size_t);
-  int (*write)(cosmic_io_t *, const char *, size_t);
+  ssize_t (*read)(cosmic_io_t *, char *, size_t);
+  ssize_t (*write)(cosmic_io_t *, const char *, size_t);
   int (*close)(cosmic_io_t *);
 };
 
@@ -10,7 +10,7 @@ struct cosmic_io {
   const struct cosmic_io_vtable *vtbl;
 };
 
-int cosmic_io_read(cosmic_io_t *io, char *buf, size_t len) {
+ssize_t cosmic_io_read(cosmic_io_t *io, char *buf, size_t len) {
   if (!io->vtbl->read) {
     return -1;
   }
@@ -18,7 +18,7 @@ int cosmic_io_read(cosmic_io_t *io, char *buf, size_t len) {
   return io->vtbl->read(io, buf, len);
 }
 
-int cosmic_io_write(cosmic_io_t *io, const char *buf, size_t len) {
+ssize_t cosmic_io_write(cosmic_io_t *io, const char *buf, size_t len) {
   if (!io->vtbl->write) {
     return -1;
   }
